Review feedback
- Adds some better documentation around the semantics of dropping a port
  ticket
- Adds links to issues in some todos
- Splits illumos/other platform implementations at module level.
