+++ sled-agent/src/opte/illumos/port.rs
@@ -4,12 +4,11 @@
 
 //! A single port on the OPTE virtual switch.
 
-use super::port_manager::PortTicket;
-use super::BoundaryServices;
-use super::Gateway;
-use super::Vni;
-#[cfg(target_os = "illumos")]
 use crate::illumos::dladm::Dladm;
+use crate::opte::BoundaryServices;
+use crate::opte::Gateway;
+use crate::opte::PortTicket;
+use crate::opte::Vni;
 use crate::params::ExternalIp;
 use ipnetwork::IpNetwork;
 use macaddr::MacAddr6;
@@ -18,7 +17,6 @@ use std::net::Ipv6Addr;
 use std::sync::Arc;
 
 #[derive(Debug)]
-#[cfg_attr(not(target_os = "illumos"), allow(dead_code))]
 struct PortInner {
     // Contains instance ID and a pointer to the parent manager
     ticket: PortTicket,
@@ -56,7 +54,6 @@ struct PortInner {
     vnic: String,
 }
 
-#[cfg(target_os = "illumos")]
 impl Drop for PortInner {
     fn drop(&mut self) {
         if let Err(e) = Dladm::delete_vnic(&self.vnic) {
